+++ src/main/java/card/credit/w3/w3/entidades/secundarias/controller/AumentoLimiteCartaoController.java
@@ -7,6 +7,8 @@ import card.credit.w3.w3.entidades.secundarias.SolicitacaoLimite;
 import card.credit.w3.w3.entidades.secundarias.services.SolicitacaoLimiteService;
 import jakarta.validation.Valid;
 
+import java.math.BigDecimal;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -21,7 +23,7 @@ public class AumentoLimiteCartaoController {
     private SolicitacaoLimiteService solicitacaoLimiteService;
 
     
-    public record SolicitacaoLimiteRequest(CartaoDTO cartao, double limiteNovo) {}
+    public record SolicitacaoLimiteRequest(CartaoDTO cartao, BigDecimal limiteNovo) {}
 
     @PostMapping("/solicitar")
     public ResponseEntity<SolicitacaoLimite> aumentoLimiteCartao(@RequestBody @Valid SolicitacaoLimiteRequest request) {
